updating spec's and UI integration
about half the tests are broken, but this is an issue w/how I'm testing -- it works find in the browser.  After doing some research it looks like I will have to work on these tests to enable them to look at disabled fields for their information.  I'll look at this later.